FinalProject/pyproject.toml
[project]
name = "cosc-520-final-project"
version = "0.1.0"
description = "Final Project for COSC 520"
authors = [
    {name = "Rishav Banerjee", email = "rishavbanerjee1710@gmail.com"},
    {name = "Zhenbang He", email = "zbhe96@student.ubc.ca"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
typeguard = "^4.4.4"
tqdm = "^4.67.1"
pytest = "^8.4.2"
mmh3 = "^5.2.0"
numpy = "^2.3.4"
pandas = "^2.3.3"
fuzzy-search = "^2.5.0"
matplotlib = "^3.10.7"
seaborn = "^0.13.2"
gutenberg = "^0.8.0"
